feat(rh): Sub-PR 5 shared-modules-refactor — empleado-detail + contrato/finiquito + RDB rollout
Cierra la iniciativa `shared-modules-refactor` con la consolidación
del detalle de empleado y la generación de contrato/finiquito como
módulos compartidos cross-empresa. RDB hereda todas las features
(beneficiarios Art. 501, documentos, notas, multi-tel, contacto
emergencia, contrato individual y convenio de finiquito) que antes
eran "solo DILESA".

Política nueva (codificada como SM6 en ADR-011): cero fallback
hardcoded en módulos que generan documentos legales. `PATRON_DILESA`
eliminado de `contrato-printable.tsx` y `finiquito-printable.tsx`. Los
datos del patrón se construyen desde `core.empresas` con el helper
centralizado `lib/rh/datos-fiscales-empresa.ts`. Sin esos datos, el
botón "Nuevo empleado" se deshabilita en single-empresa y los botones
"Contrato"/"Finiquito" se deshabilitan en el detalle. Las pages
`/contrato` y `/finiquito` muestran mensaje claro con CTA a
Settings → Empresas si alguien llega por URL directa.

DILESA mantiene operación normal (datos fiscales completos en BSOP).
RDB queda pendiente de capturar CSF para destrabar uso real de RH
formal — sin riesgo de regresión: el detail page sigue funcionando,
solo contrato/finiquito y altas nuevas requieren los datos.

- `lib/rh/datos-fiscales-empresa.ts` — validador centralizado
  (`camposFaltantes`, `tieneDatosCompletos`, `buildPatronFromDatos`,
  `useDatosFiscalesEmpresa`).
- `components/rh/empleado-detail-module.tsx` (~1100 líneas) adoptando
  DILESA como base canónica.
- `components/rh/empleado-contrato-module.tsx` y
  `components/rh/empleado-finiquito-module.tsx` con bloqueo duro si
  faltan datos.
- Pages reducidos:
  - DILESA `/personal/[id]/page.tsx` 1441 → 5 líneas.
  - RDB    `/personal/[id]/page.tsx`  679 → 5 líneas.
  - DILESA `/personal/[id]/contrato/page.tsx` 232 → 5 líneas.
  - DILESA `/personal/[id]/finiquito/page.tsx` 353 → 5 líneas.
- Rutas espejo creadas:
  - RDB `/personal/[id]/contrato/page.tsx`.
  - RDB `/personal/[id]/finiquito/page.tsx`.
- `personal-module.tsx`: guard de alta deshabilita el botón cuando la
  empresa no tiene datos fiscales o cuando el scope es multi-empresa.
- ADR-011 actualizado con regla SM6 + ajuste a SM4 documentando la
  alternativa de "escalar la feature" cuando aplique a todas.
- `INITIATIVES.md`: `shared-modules-refactor` movida a Done;
  `forms-pattern` queda libre como siguiente en cola UI.
