+++ app/rdb/rh/personal/[id]/contrato/page.tsx
@@ -1,5 +1,5 @@
 import { EmpleadoContratoModule } from '@/components/rh/empleado-contrato-module';
 
 export default function Page() {
-  return <EmpleadoContratoModule empresaSlug="dilesa" />;
+  return <EmpleadoContratoModule empresaSlug="rdb" />;
 }
