+++ app/rdb/rh/personal/[id]/finiquito/page.tsx
@@ -1,5 +1,5 @@
 import { EmpleadoFiniquitoModule } from '@/components/rh/empleado-finiquito-module';
 
 export default function Page() {
-  return <EmpleadoFiniquitoModule empresaSlug="dilesa" />;
+  return <EmpleadoFiniquitoModule empresaSlug="rdb" />;
 }
